common grouping mechanism for courts
