+++ peachjam/templates/peachjam/_document_table_row.html
@@ -26,6 +26,7 @@
           <i class="bi bi-translate"></i>
         </span>
       {% endif %}
+      {% include 'peachjam/_labels.html' with labels=document.labels.all %}
     </td>
   {% endif %}
   {% if doc_table_citations %}<td class="cell-citation">{{ document.citation|default_if_none:'' }}</td>{% endif %}
